+++ nitorch/io/babel/__init__.py
@@ -1,0 +1,3 @@
+from .array import BabelArray
+
+from . import array, metadata, utils
